I have worked on this commit for way too long but here it is

+++ Engine/Renderer/OpenGL/Components/GLMesh.cpp
@@ -1,21 +1,21 @@
 // GLMesh.cpp
-// Last update 3/27/2021 by Madman10K
+// Last update 4/12/2021 by Madman10K
 #include "GLMesh.hpp"
 
 void UVK::GLMesh::createMesh(GLfloat* vertices, uint32_t* indices, uint32_t vertexNum, uint32_t indexNum)
 {
-	indexCount = indexNum;
+	indexCount = static_cast<int>(indexNum);
 
 	glGenVertexArrays(1, &VAO);
 	glBindVertexArray(VAO);
 
 	glGenBuffers(1, &IBO);
 	glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, IBO);
-	glBufferData(GL_ELEMENT_ARRAY_BUFFER, sizeof(indices[0]) * indexNum, indices, GL_STATIC_DRAW);
+	glBufferData(GL_ELEMENT_ARRAY_BUFFER, static_cast<long int>(sizeof(indices[0])) * indexNum, indices, GL_STATIC_DRAW);
 
 	glGenBuffers(1, &VBO);
 	glBindBuffer(GL_ARRAY_BUFFER, VBO);
-	glBufferData(GL_ARRAY_BUFFER, sizeof(vertices[0]) * vertexNum, vertices, GL_STATIC_DRAW);
+	glBufferData(GL_ARRAY_BUFFER, static_cast<long int>(sizeof(vertices[0])) * vertexNum, vertices, GL_STATIC_DRAW);
 
 	glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, sizeof(vertices[0]) * 5, nullptr);
 	glEnableVertexAttribArray(0);
